Allow label sets initialization while registering metrics
`init_label_sets` is a new config that expects a Hash with metrics names
as keys and an array of label sets as values.

Example:

```ruby
{
  metric_name: [{label: 'value1', other_label: 'value1'}, {label: 'value1', other_label: 'value2'}],
  another_metric_name: [{label: 'value1', other_label: 'value1'}]
}
```

Label sets are fetched during metric registration and initialized by calling
[metric.init_label_set](https://github.com/prometheus/client_ruby#init_label_set).

This is specially useful for initializing global metrics since they are registered
upon leader election.